Refactor: improves code organization
