--- temlogger/formatter.py
@@ -1,62 +1,0 @@
-from logstash.formatter import LogstashFormatterBase
-
-
-class FormatterBase(LogstashFormatterBase):
-
-    def __init__(self, tags=None, fqdn=False, environment='', **kwargs):
-        super().__init__(message_type='', tags=None, fqdn=False, **kwargs)
-        self.environment = environment
-
-
-class StackDriverFormatter(FormatterBase):
-
-    def format(self, record):
-        # Create message dict
-        message = {
-            '@timestamp': self.format_timestamp(record.created),
-            'message': record.getMessage(),
-            'host': self.host,
-            'path': record.pathname,
-            'tags': self.tags,
-            'environment': self.environment,
-
-            # Extra Fields
-            'level': record.levelname,
-            'logger_name': record.name,
-        }
-
-        # Add extra fields
-        message.update(self.get_extra_fields(record))
-
-        # If exception, add debug info
-        if record.exc_info:
-            message.update(self.get_debug_fields(record))
-
-        return message
-
-
-class LogstashFormatter(FormatterBase):
-
-    def format(self, record):
-        # Create message dict
-        message = {
-            '@timestamp': self.format_timestamp(record.created),
-            'message': record.getMessage(),
-            'host': self.host,
-            'path': record.pathname,
-            'tags': self.tags,
-            'environment': self.environment,
-
-            # Extra Fields
-            'level': record.levelname,
-            'logger_name': record.name,
-        }
-
-        # Add extra fields
-        message.update(self.get_extra_fields(record))
-
-        # If exception, add debug info
-        if record.exc_info:
-            message.update(self.get_debug_fields(record))
-
-        return self.serialize(message)
